requirements.txt
amqp==1.4.9
aniso8601==1.2.0
anyjson==0.3.3
appdirs==1.4.0
Authomatic==0.0.13
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.4.0
billiard==3.3.0.23
boto==2.45.0
bz2file==0.98
celery==3.1.18
click==6.6
coverage==3.6
cycler==0.10.0
decorator==4.0.2
Django==1.8.2
django-dotenv==1.4.1
django-filter==0.11.0
django-pgjson==0.2.3
djangorestframework==3.2.4
ecdsa==0.13
enum34==1.1.6
et-xmlfile==1.0.1
feedparser==5.2.1
functools32==3.2.3.post2
funcy==1.5
future==0.14.3
futures==3.0.3
gensim==0.11.1
google-api-python-client==1.4.2
gunicorn==19.3.0
httplib2==0.9.2
igraph==0.1.11
ipython==5.1.0
ipython-genutils==0.1.0
isodate==0.5.4
jdcal==1.3
jgraph==0.2.1
Jinja2==2.7.2
joblib==0.8.4
jsonschema==2.4.0
kombu==3.0.37
lxml==3.4.4
MarkupSafe==0.23
matplotlib==2.0.0
natsort==5.0.2
networkx==1.10
nltk==3.0.4
nose==1.3.7
numexpr==2.4.3
numpy==1.12.0
oauth2==1.9.0.post1
oauth2client==1.5.1
oauthlib==0.7.2
openpyxl==2.3.0
packaging==16.8
pandas==0.19.2
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.8
psycopg2==2.6
ptyprocess==0.5.1
py-trello==0.8.0
pyasn1==0.2.1
pyasn1-modules==0.0.8
pycrypto==2.6.1
PyGithub==1.32
Pygments==2.1.3
pyLDAvis==1.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
python-dotenv==0.6.0
python-jose==1.3.2
pytz==2016.10
rdflib==4.2.2
redis==2.10.3
requests==2.7.0
requests-oauthlib==0.8.0
rsa==3.4.2
scikit-bio==0.2.3
scikit-learn==0.16.1
scipy==0.15.1
simplegeneric==0.8.1
six==1.10.0
smart-open==1.3.5
subprocess32==3.2.7
tincan==0.0.5
traitlets==4.3.1
twython==3.2.0
uritemplate==3.0.0
vaderSentiment==0.5
wcwidth==0.1.7
